models: db: implement the document store middleware
Implement the middleware that handles requests towards the document
database. This database stores the measurements and sensors. Controllers
and background services use the IMeasurementInterface to read and write
data from the database. Measurement model validation / building is done
by the datalayer, since a measurement can have so many points of origin
(MQTT, websockets, REST call, etc..).

Measurements are received in a strict format, implemented by the
`RawMeasurement' class, implemented by the AbstractMeasurementRepository
as an internal class.

Signed-off-by: Michel Megens <dev@bietje.net>