infra: measurements: throw an exception on validation faults
If `AbstractMeasurementRepository.StoreMeasurement()' runs into a validation
fault, an InvalidRequestException is thrown. The exception error code is
set according to the reason of the throw:

	* Error code 300 is set for JSON errors
	* Error code 301 is set for secret errors

Both errors are kept purposefully generic; the less the client knows the better.
Websockets might also return a status code of 500. This code indicates (like
HTTP) an internal (database) server error. This status code is returned when, for
example, a DatabaseException is thrown. A status code of 200 means "OK".

Signed-off-by: Michel Megens <dev@bietje.net>