infra: measurements: simplify class design
Simplify the class design / inheritance tree of the measurement
repositories. The `AbstractMeasurementRepository' was an unnecessary
abstraction layer. This functionality is now implemented in the
`MeasurementRepository', which is the MongoDB abstraction layer. The cache
is still implemented in the `CachedMeasurementRepository', which inherits
from `MeasurementRepository'.

The design of the measurement model is also fundamentally different. The
model now stores an arbitrary JSON object under the 'Data' attribute. It is
up to the client to define the layout of this object.

Signed-off-by: Michel Megens <dev@bietje.net>