infra: cache: implement absolute cache expiration
Implement the posibility to use an absolute expiry timestamp on (distributed)
cache entries. This is a use feature for methods that use return joined
database records. Currently the absolute expiry is only used by the `GetAfter()'
and `GetAfterAsync()' methods from in the measurements repository.

Cache entries created by the methods mentioned above will persist for a
maximum of 4 minutes, regardless of the number of lookups. Other methods
don't need the absolute expiry as their results won't change with the same
input figures.

Signed-off-by: Michel Megens <dev@bietje.net>