infra: measurements: validate received data
Validate the data received in a measurement using `BsonDocument.TryParse()'. If
this method fails, then the data that was pushed to Sensate isn't in
proper JSON format and can therefore not be stored safely in MongoDB. An
exception will be thrown (InvalidRequestException) to handle such an
error. The exception is to be handled by the middleware that received
the measurement (MQTT or Websocket).

Doing said validation here allows us to to disable the document
validation by MongoDB through the insert options. This reduces the load
on the MongoDB servers significantly.

Signed-off-by: Michel Megens <dev@bietje.net>