core: models: move classes to the correct namespace
Move classes in the middleware namespace to the generic models
namespace, as this better describes their functionality. The live data
handling has been removed from the websocket handler project. This
feature is to be implemented in a separate microservice.

Remove live data functionality from the websocket handler project. This
feature is to be implemented in a seperate microservice.

Signed-off-by: Michel Megens <dev@bietje.net>