infra: storage: improve processing throughput
The measurement processing implementation has the potential to create
high amounts of I/O-bound operations. The user roles are lookup
individually and then cached locally. This is not a problem as long
as there are not too many different user accounts involved.

The new processing implementation uses an asynchronous operation to
lookup all required user and its roles before processing begins. The
throughput with few users stays the same, while the throughput with
many different users involved improves dramatically. Users and user
roles are cached locally for up to 1 hour.

In order to improve read performance, measurements and audit logs are
inserted in a descending order based on their creation timestamp.

Signed-off-by: Michel Megens <dev@bietje.net>