infra: storage: prevent empty/invalid database calls
If all measurements in the processing queue are invalid, the measurement
cache will attempt to store an empty array of measurements, as well as
incorrectly incrementing the sensor statistics and invoking the cache's
event handlers.

The (cached) measurement store will always attempt to create an audit
log entry. However, if MongoDB throws an exception due to, for example,
an empty insert queue, the insertion of audit logs in also cancelled using
cancellation tokens. To prevent empty database calls, the database calls
have been wrapped into an `if' clause, together with the event calls.

Always invoked:

	* Audit log storage.

Invoked if measurement queue > 0

	* Storage of measurements;
	* Statistics increment;
	* Event calls.

Signed-off-by: Michel Megens <dev@bietje.net>