infra: storage: improve measurement cache iops
Improve the measurement cache performance in serveral ways. I/O
operations should be started as soon as possible (asynchronously). The
cache processor starts several I/O operations assynchronously and the
creation of audit logs can be done in parallel with raw measurement
parsing. This reduces the load on the database server, by spreading
different request over a greater timeframe.

Update the MongoDB insert options and write concern for measurements
and audit log entries. The document structure is validated entirely by
the application and therefore the database server doesn't have to do
it again. This reduces the load on the database server.

Lock contention isn't an issue in the measurement cache implementation.
Therefore, the load on the database is further reduced by implementing
the measurement cache as a singleton.

Signed-off-by: Michel Megens <dev@bietje.net>