infra: storage: optimze measurement storage
Optimize the storage of measurements, by improving both the storage size
and storage speed per measurement. The measurement store no longer stores
individual measurements, but buckets of measurements. Doing so improves the
storage size by more than 50%. The measurement collection is indexed on
the following properties:

	* Internal ID: _id
	* Sensor ID: SensorId
	* Timestamp (hour): Timestamp
	* Measurement count: Count
	* First measurement timestamp: First
	* Last measurement timestamp: Last

The measurement collection itself is not indexed. Measurements are appended
to a bucket as long as the bucket size is smaller than 500 and all
measurements are taken in the same hour.

In order to reduce the database load, measurement stores no longer creates
audit log entries. Instead, the request method is now stored on the statistics
entry. The request method has been added to the indexes of the statistics
collection.

Signed-off-by: Michel Megens <dev@bietje.net>