infra: measurements: update measurement structure
Update the measurement storage structure. The data points are no longer
stored in an array, but in a dictionary instead. This datastructure
makes retrieving data on clients much easier.

The location (longitude and latitude) are no longer required elements
of a measurement. The longitude and latitude can be added to the data
dictionary.

Signed-off-by: Michel Megens <dev@bietje.net>