infra: storage: optimze the cached measurement store
In order to reduce the spike load on MongoDB and therefore improve the
(cached) measurement store performance, I/O operations are started as
soon as possible. This means that the incrementing of the statistics can
be started way before the actual storage of measurements can be started.
In an ideal situation the statistics query finished asynchronously before
the measurement storage operation is started.

Performance can be further improved by improving the measurement validation
process. The measurement model is manually validated by the storage engines
implemented in the `Infrastructure.Storage' subsystem. The validation
checks done by the engines are now ordered from simple (cheap) to complex
(expensive) whenever possible. This optimization might save some CPU time
if the simple validations fail first.

Improve the storage size of the measurement buckets. The measurement buckets
have been reduced in size by removing the `First' and `Last' attributes.
These attributes store the first and last measurement timestamp within the
measurement bucket respectively. The idea behind storing these values is the
(potential) improvement on aggragate queries. In practical situations this
improvement is neglectible or non existent.

Signed-off-by: Michel Megens <michel@michelmegens.net>