infra: storage: improve measurement authorization
Implement an additional method to authorize measurements in the
ingress data services. The "old" (but still supported) method of
authorization is sending a plain text API key with every
measurement. The new method utilizes a SHA256 signature, by checking
the following condition:

sha256(measurement) == secret

Where `secret' is the received secret, `measurement' the received
measurement in plain text and `sha256()' is the SHA256 hashing
algorithm.

The client device/smart sensors create a `RawMeasurement' object as
usual, including the API key. Before sending said measurement, the
device computes the SHA256 hash of the entire measurement and stores
the SHA256 hash in the place of the API/secret. When the platform
receives the measurement the process will be reversed.

The platform checks if the `CreatedBySecret' attribute of the
incoming measurement is an API key (old flow) or SHA256 hash. If
the latter is the case, the platform attempts to compute the same
hash using the secret found in the sensor database. If the hashes
match, the measurement is correctly authorized.

Signed-off-by: Michel Megens <michel@michelmegens.net>