auth-service: cache: implement a custom cache backend
Implement a red-black tree to serve as the backend of serveral
data caches. This implementation is optimized for fast, multi-core,
lookups, while maintaining decent insert and delete performance.

The performance gains are archieved by supporting parallelism
natively. A read-write lock has been incorperated in order to prevent
data races. This means that all search/lookup operations can be
computed in parallel. Insert and delete operations are always sequential.

The second feature that is unique to this implementation is native support
for timeouts using `boost::chrono'. This way the life time of objects
can be controlled during insertion. Lookups that target an object that
is no longer valid will never be returned by `RawFind()'. Instead, the
implementation acts as if no object was found.

Signed-off-by: Michel Megens <michel@michelmegens.net>