common: caching: implement a read optimized cache
Implement an in-memory cache optimized for read operations. The
implementation is fully thread-safe, but utilizes the synchronous data
structures provided by the .NET runtime (Dictionary<K, V>). This path
is chosen in order to implement bulk insertions and deletions. These bulk
operations improve cache population times drastically.

Users of the cache don't need to keep track of what has been added,
because the cache will do so by adding a timeout. After a given period
an entry will automatically be marked for removal. The actual removal
of entry's (because it was activly removed, or due to a timeout) happens
asynchronously on a worker thread.

Signed-off-by: Michel Megens <michel@michelmegens.net>