auth-service: caching: implement a read optimized cache
Implement an in-memory cache optimized for read operations. The
implementation is fully thread-safe, but utilizes the synchronous data
structures provided by a third-party (a hopscotch hash table).
Implementing our own hash table or tree-map is not worth the time for
the (minimal) time savings possible. This memory cache supports bulk
operations in order to drastically improve the cache polulation
performance.

Users of the cache don't need to keep track of what has been added,
because the cache will do so by adding a timeout. After a given period
an entry will automatically be marked for removal. The actual removal of
an element can be realized through a call to `Cleanup()'. This method
will remove all expired entries.

Signed-off-by: Michel Megens <michel@michelmegens.net>