network: queueing: implement a concurrent queue
Implement a deque datastructure. This implementation forms the backend
of the message queue. A deque is a queue implemented as a list. This means
that the following operations are all O(1):

	* Enqueue to the front or back;
	* Dequeue from the front or back;
	* Random lookup.

The random lookups are important in the message queue to simplify and
optimize statistical/metric calculations. This deque is implemented with
concurrency in mind, by protecting all public operations with a spinlock.

Signed-off-by: Michel Megens <michel@michelmegens.net>