network: routing: implement a routing cache
Implement an in-memory cache optimized for read operations. This cache
is used to cache the information needed by the message router. The
implementation is fully thread-safe, but utilizes the synchronous data
structures provided by the .NET runtime (Dictionary<K, V>). By caching
data locally it is much easier to provision the cache using bulk insert
operations.

Implementators of the cache can load the cache using the `Load' methods.
These methods will (re)load the data into the cache, resetting any data
that was previously stored. Only the data of the respective load is
reset.

Signed-off-by: Michel Megens <michel@michelmegens.net>