routing: implement a new routing subsystem
Implement a routing subsystem based on the composite design pattern. This
pattern is used to improve the ability to add other routing rules in the
future. This rule engine allows implementing routing logic in seperated
code units.

The composite router is used to execute all registered rules, in
parallel. The order of router execution is determined by the order in
which routers (rules) are added to the composite router.

Signed-off-by: Michel Megens <michel@michelmegens.net>