routers: composite-router: improve parallelism
Improve the parallelism of the composite router. The interface
of the composite router now accepts a batch of messages that will
be processed using the TPL subsystem. Incorrect messages are
dropped, and only external problems (such as unreachable databases
or caches) will cause the router to drop an entire batch.

Signed-off-by: Michel Megens <michel@michelmegens.net>