routing: cache: improve routing cache implementation
Improve the routing cache implementation by splitting several
large methods into smaller chunks. This improves the maintainability
of the the entire cache, but specificly the live data flow.

Further more, errors in the routing cache are logged on the appropriate
logging level. Until now, the routing cache only logged on the 'debug'
level. Warnings are now raised to the 'warning' level.

Signed-off-by: Michel Megens <michel@michelmegens.net>