routing: composite-router: improve routing code
Improve the overall code quality of the composite router. This is
archieved by moving authorization code into its own router: the
Authorization Router. This way the composite router is clear of
business logic, except for the management and execution process of
different routers. The only validation the composite router still
does, is validating that sensor, account ID and sensor key contain
valid values.

Signed-off-by: Michel Megens <michel@michelmegens.net>