routing: composite-router: improve logging
Log dropped messages on the network event and allow tracking of dropped
messages on a per sensor basis. This improves the self service level of
customers, because failing sensors are easier to spot.

Improve logging in the composite router to be able to track issues in
specific routers. The composite router now logs which router is being
called before calling said router. To prevent clutter, this is logged on
the `Debug` log level. The log level to trace unknown sensors has been
incremented from `Debug` to `Warning`.

An error log is written if an uncaught exception is thrown in the
parallel routing operation. The inner exception of the aggregate
exception is rethrown after logging the complete trace.

Signed-off-by: Michel Megens <michel@michelmegens.net>