routing: processing: improve message processing
Improve the message processing in the composite router. The
composite router is made more autonomous by migrating the
dequeuing of messages from the routing service to the composite
router. This way the router can process messages more efficiently
and is the routing service only responsible for the timing of the
routing process. The routing process is started by calling the
`TryRoute()` method on the `IMessageRouter` interface.

The routing unit tests have been adapted to test this new
behavior.

Signed-off-by: Michel Megens <michel@michelmegens.net>